+++ src/App.js
@@ -3,7 +3,7 @@ import './App.css';
 import { Routes, Route } from 'react-router-dom';
 import HoldingsDetail from './Holdings/HoldingsDetail.js';
 import PortfolioList from './Portfolio/PortfolioList';
-import TransactionsList from './Holdings/TransactionsList';
+import TransactionsList from './Transactions/TransactionsList';
 import Dividends from './Dividends/Dividends';
